refactor(queue): structurally shrink giant modules
